requirements.txt
backcall==0.2.0
Box2D==2.3.10
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
glcontext==2.3.6
imageio==2.22.1
ipykernel==6.16.0
ipython==7.34.0
jedi==0.18.1
jupyter-client==7.3.5
jupyter-core==4.11.1
matplotlib-inline==0.1.6
moderngl==5.6.4
moderngl-window==2.4.2
multipledispatch==0.6.0
nest-asyncio==1.5.6
numpy==1.21.6
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
psutil==5.9.2
pygame==2.1.2
pyglet==2.0a4
Pygments==2.13.0
pyparsing==3.0.9
pyrr==0.10.3
python-dateutil==2.8.2
PyWavefront==1.3.3
pywin32==304
pyzmq==24.0.1
six==1.16.0
snakeviz==2.1.1
tornado==6.2
traitlets==5.4.0
wcwidth==0.2.5
